HttpExample/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-cognitiveservices-vision-customvision
azure-common
certifi
chardet
click
Flask
future
idna
isodate
itsdangerous
Jinja2
MarkupSafe
msrest
numpy
oauthlib
opencv-python
pyngrok
PyYAML
requests
requests-oauthlib
six
urllib3
Werkzeug
